Crate the_process_foundry

The root library of the Process Foundry

This is here to bootstrap and tie the entire system together.

TODO: Make more of a README for this

Some Definitions

Application - An item which can run actions and generate events

Container - An abstract object which contains applications and other containers. It can be used to look

its own contents up and pass actions to them.

Module - An implementation of an interface to an Application

Registry - This a singleton that maintains a searchable list of application and container factories

Workflow - A set of Applications wired together to implement a single process

Modules

applications
A placeholder to group all the developmental applications

base
The core definitions and structs for the Process Foundry

error
Errors that can be generated by the Process Foundry

Structs

Rc
A single-threaded reference-counting pointer. 'Rc' stands for 'Reference Counted'.

VolumeQuery

Functions

bootstrap

find

get_or_create_volume

main