Module the_process_foundry::application

A wrapper around external software to generate TPF Actions and Events

An application (app) is an interface which defines a set of functions to convert TPF actions into native commands, monitor state, and sends out messages describing any changes in state.

At its core, an application is something that executes actions and emits events. It has no context of anything outside the scope of its functionality. It can be accessed by one or more containers, depending on the context.

For more detailed info about the architecture, read https://lurkingfrog.github.io/the_process_foundry_book/core/application.html

Structs

ApiAccess

AppDescription
General information about the external application handled by this particular module

AppInstance
Information about the specific running copy of the external program

CliAccess

Traits

AppTrait
A data structure containing information used to generate both generate Actions and Events