Module the_process_foundry::applications::bash

Use and manage Bash shell

This is essentially my hello world application and using it as a stalking horse for finding user stories THINK: Considering to change this to a generic unix shell with a plugin for specific type (Bash, ZSH), since they seem to be common with the differences coming from the scripting and interactive portions

Structs

Bash

FindAppQuery
Configuration to look up an application in this container TODO: Add a macro to map all the functions to the parent AppQuery

RunOptions

RunResult

Enums

Action

ActionResult

Constants

APP_NAME

MODULE_VERSION