Module the_process_foundry::applications::docker_compose::schema

Schema for a docker-compose configuration file

This is the data structure for holding the yaml file that docker compose uses to do its thing. I'm only doing the fields for features I'm using as a proof of concept. This schema structure is only going to be explicitly used by the application, so it should all be private and accessed by getters/setters.

TODO: Write stand alone CLI tool to convert JSON to structs "schema.rs" or some such. TODO: Write CLI tool to diff schema versions TOOD: Figure out versioning (#[serde_semver(x)], where x is semver::VersionReq. https://docs.rs/semver/0.10.0/semver/struct.VersionReq.html)

Structs

Build

PortObj

Schema
A Schema structure to contain all of the possible values that can be contained in a docker-compose.yml This is going to be incomplete, only adding things as I implement functions. See https://github.com/docker/compose/tree/master/compose/config for the JSON definitions of this.

Service

ServiceVolume

ServiceVolumeTmpfs

Enums

Port

Restart

ServiceVolumeConsistency

ServiceVolumeType