Module the_process_foundry::applications::pg_basebackup

A wrapper for the CLI tool pg_basebackup

This is used to generate backup files from a running Postgres database https://www.postgresql.org/docs/12/app-pgbasebackup.html

Structs

Options
All the command line options that can be passed to the program

PgBaseBackup

Enums

Action

ActionResult

Checkpoint

Compression
The encoding of the output file

Rate

WalMethod

Constants

APP_NAME

MODULE_VERSION