Module the_process_foundry::base

The core definitions and structs for the Process Foundry

After getting lost in the weeds, this is all the things that I seem to see come up as useful in every application and container

THINK: Additional traits

Actionable: Possibly part of app trait, since all should be able to utilize and emit Actions/Events

ActionResult: To abstract the result so we can pass it to something that has the code to actually use

Routable (possibly part of action)

consider https://abronan.com/rust-trait-objects-box-and-rc/ - Arc<Mutex>

Structs

ApiAccess

AppInstance
A description of all the known information about discovered application.

AppQuery
This defines the a needed version of container/application needed for the module.

CliAccess

CliQueryOptions

Cmd

Shell
A special case for bootstrapping. I'm trying to find the enumerations that actually deserve to be traits themselves

Enums

Message
A generic message designed to be sent to a container

QueryOptions
A wrapper for looking for specific instances of apps or where they are running

ShellType

Traits

ActionTrait
Handlers for serialized action requests

AppTrait
All applications should be introspective and able to build a running manager of itself

ContainerTrait
Ways to manage applications (eg Docker, Bash) contained within itself

LocalTrait
An app running on the local system (eg: bash shell)