Trait the_process_foundry::base::AppTrait

pub trait AppTrait {
    fn build(
        instance: AppInstance, 
        parent: Option<Rc<dyn ContainerTrait>>
    ) -> Result<Self>
    where
        Self: Sized;

    fn get_name(&self) -> String;

    fn set_version(&self, instance: AppInstance) -> Result<AppInstance>;

    fn set_cli(
        &self, 
        instance: AppInstance, 
        container: Rc<dyn ContainerTrait>
    ) -> Result<AppInstance>;
}

All applications should be introspective and able to build a running manager of itself

Required methods

fn build(
instance: AppInstance,
parent: Option<Rc<dyn ContainerTrait>>
) -> Result<Self> where
Self: Sized,

Construct the metadata for the module controlling the app instance This is important to make sure the module aligns with the actual installed instance

fn get_name(&self) -> String

Print a standardized "name (version)" for logging purposes

fn set_version(&self, instance: AppInstance) -> Result<AppInstance>

Knows how to get the version number of the installed app (not the module version)

fn set_cli(
&self,
instance: AppInstance,
container: Rc<dyn ContainerTrait>
) -> Result<AppInstance>

Figures out how to call the cli using the given container THINK: is it better to have an option or make "Local" a special case?

Loading content...