Trait the_process_foundry::base::ContainerTrait

pub trait ContainerTrait: Debug {
    fn find(&self, query: AppQuery) -> Result<Vec<AppInstance>>;

    fn forward(&self, to: AppInstance, message: Message) -> Result<String>;

    fn cached_apps(&self) -> Result<Vec<AppInstance>>;

    fn get_name(&self) -> String;

    fn find_one(&self, query: AppQuery) -> Result<AppInstance> { ... }
}

Ways to manage applications (eg Docker, Bash) contained within itself

Required methods

fn find(&self, query: AppQuery) -> Result<Vec<AppInstance>>

This will find a list of apps with configurations that the container knows about

fn forward(&self, to: AppInstance, message: Message) -> Result<String>

Send a stringified action to the AppInstance

fn cached_apps(&self) -> Result<Vec<AppInstance>>

List the known items in the app cache

fn get_name(&self) -> String

Get the name/version of the container, usually for use in logging/errors.

Loading content...
Provided methods

fn find_one(&self, query: AppQuery) -> Result<AppInstance>

Find a unique app that matches the query

Loading content...